# Implement histogram iterators for batch