--- a/pkg/querier/batch/batch.go
+++ b/pkg/querier/batch/batch.go
@@ -31,10 +31,10 @@ func (c GenericChunk) Iterator(reuse chunk.Iterator) chunk.Iterator {
 
 // iterator iterates over batches.
 type iterator interface {
-	// Seek to the batch at (or after) time t.
+	// Seek to the batch at (or after) time t and returns chunk value type.
 	Seek(t int64, size int) chunkenc.ValueType
 
-	// Next moves to the next batch.
+	// Next moves to the next batch and returns chunk value type.
 	Next(size int) chunkenc.ValueType
 
 	// AtTime returns the start time of the next batch.  Must only be called after
@@ -44,7 +44,7 @@ type iterator interface {
 	// MaxCurrentChunkTime returns the max time on the current chunk.
 	MaxCurrentChunkTime() int64
 
-	// Batch returns the current batch.  Must only be called after Seek or Next
+	// Batch returns the current batch. Must only be called after Seek or Next
 	// have returned true.
 	Batch() chunk.Batch
 
@@ -157,12 +157,12 @@ func (a *iteratorAdapter) Err() error {
 
 // AtHistogram implements chunkenc.Iterator.
 func (a *iteratorAdapter) AtHistogram(h *histogram.Histogram) (int64, *histogram.Histogram) {
-	return a.curr.Timestamps[a.curr.Index], a.curr.Histograms[a.curr.Index]
+	return a.curr.Timestamps[a.curr.Index], (*histogram.Histogram)(a.curr.HistogramValues[a.curr.Index])
 }
 
 // AtFloatHistogram implements chunkenc.Iterator.
 func (a *iteratorAdapter) AtFloatHistogram(h *histogram.FloatHistogram) (int64, *histogram.FloatHistogram) {
-	return a.curr.Timestamps[a.curr.Index], a.curr.FloatHistograms[a.curr.Index]
+	return a.curr.Timestamps[a.curr.Index], (*histogram.FloatHistogram)(a.curr.HistogramValues[a.curr.Index])
 }
 
 // AtT implements chunkenc.Iterator.