--- a/pkg/querier/batch/stream.go
+++ b/pkg/querier/batch/stream.go
@@ -1,8 +1,6 @@
 package batch
 
 import (
-	"unsafe"
-
 	"github.com/prometheus/prometheus/model/histogram"
 	"github.com/prometheus/prometheus/tsdb/chunkenc"
 
@@ -38,12 +36,12 @@ func (bs *batchStream) next() {
 
 func (bs *batchStream) atHistogram() (int64, *histogram.Histogram) {
 	b := &(*bs)[0]
-	return b.Timestamps[b.Index], (*histogram.Histogram)(b.HistogramValues[b.Index])
+	return b.Timestamps[b.Index], b.Histograms[b.Index]
 }
 
 func (bs *batchStream) atFloatHistogram() (int64, *histogram.FloatHistogram) {
 	b := &(*bs)[0]
-	return b.Timestamps[b.Index], (*histogram.FloatHistogram)(b.HistogramValues[b.Index])
+	return b.Timestamps[b.Index], b.FloatHistograms[b.Index]
 }
 
 func (bs *batchStream) atTime() int64 {
@@ -91,13 +89,9 @@ func mergeStreams(left, right batchStream, result batchStream, size int) batchSt
 		case chunkenc.ValFloat:
 			b.Timestamps[b.Index], b.Values[b.Index] = bs.at()
 		case chunkenc.ValHistogram:
-			t, v := bs.atHistogram()
-			b.Timestamps[b.Index] = t
-			b.HistogramValues[b.Index] = unsafe.Pointer(v)
+			b.Timestamps[b.Index], b.Histograms[b.Index] = bs.atHistogram()
 		case chunkenc.ValFloatHistogram:
-			t, v := bs.atFloatHistogram()
-			b.Timestamps[b.Index] = t
-			b.HistogramValues[b.Index] = unsafe.Pointer(v)
+			b.Timestamps[b.Index], b.FloatHistograms[b.Index] = bs.atFloatHistogram()
 		default:
 			panic("unsupported value type")
 		}