… pkg/querier/batch/batch_test.go
… pkg/querier/batch/merge_test.go
… pkg/querier/batch/non_overlapping_test.go
… pkg/querier/batch/stream_test.go
… pkg/util/histogram/testutils.go