@yeya24 2024-05-15
Benchmark results on current implementation compared with the old code. Old version seems having much better CPU time probably since the logic is simpler. No need to check different chunk value types. It is hard to tell if CPU increase is relevant to the two additional arrays in batch type.

Old version has better allocation, too, which is expected. Although the change is small.

```
name                                                                                                                              old time/op    new time/op    delta
NewChunkMergeIterator_CreateAndIterate/chunks:_1000_samples_per_chunk:_100_duplication_factor:_1_encoding:_PrometheusXorChunk-10    2.87ms ± 3%    3.24ms ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_CreateAndIterate/chunks:_1000_samples_per_chunk:_100_duplication_factor:_3_encoding:_PrometheusXorChunk-10    8.45ms ± 0%    9.53ms ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_CreateAndIterate/chunks:_100_samples_per_chunk:_100_duplication_factor:_1_encoding:_PrometheusXorChunk-10      294µs ± 1%     327µs ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_CreateAndIterate/chunks:_100_samples_per_chunk:_100_duplication_factor:_3_encoding:_PrometheusXorChunk-10      861µs ± 2%     974µs ± 3%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_CreateAndIterate/chunks:_1_samples_per_chunk:_100_duplication_factor:_1_encoding:_PrometheusXorChunk-10       3.77µs ± 2%    4.49µs ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_CreateAndIterate/chunks:_1_samples_per_chunk:_100_duplication_factor:_3_encoding:_PrometheusXorChunk-10       10.8µs ± 2%    12.4µs ± 0%   ~     (p=0.100 n=3+3)

name                                                                                                                              old alloc/op   new alloc/op   delta
NewChunkMergeIterator_CreateAndIterate/chunks:_1000_samples_per_chunk:_100_duplication_factor:_1_encoding:_PrometheusXorChunk-10     131kB ± 0%     132kB ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_CreateAndIterate/chunks:_1000_samples_per_chunk:_100_duplication_factor:_3_encoding:_PrometheusXorChunk-10     457kB ± 0%     459kB ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_CreateAndIterate/chunks:_100_samples_per_chunk:_100_duplication_factor:_1_encoding:_PrometheusXorChunk-10     14.9kB ± 0%    15.9kB ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_CreateAndIterate/chunks:_100_samples_per_chunk:_100_duplication_factor:_3_encoding:_PrometheusXorChunk-10     50.7kB ± 0%    52.9kB ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_CreateAndIterate/chunks:_1_samples_per_chunk:_100_duplication_factor:_1_encoding:_PrometheusXorChunk-10       1.67kB ± 0%    2.68kB ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_CreateAndIterate/chunks:_1_samples_per_chunk:_100_duplication_factor:_3_encoding:_PrometheusXorChunk-10       4.01kB ± 0%    6.17kB ± 0%   ~     (p=0.100 n=3+3)

name                                                                                                                              old allocs/op  new allocs/op  delta
NewChunkMergeIterator_CreateAndIterate/chunks:_1000_samples_per_chunk:_100_duplication_factor:_1_encoding:_PrometheusXorChunk-10     2.01k ± 0%     2.01k ± 0%   ~     (all equal)
NewChunkMergeIterator_CreateAndIterate/chunks:_1000_samples_per_chunk:_100_duplication_factor:_3_encoding:_PrometheusXorChunk-10     6.02k ± 0%     6.02k ± 0%   ~     (all equal)
NewChunkMergeIterator_CreateAndIterate/chunks:_100_samples_per_chunk:_100_duplication_factor:_1_encoding:_PrometheusXorChunk-10        213 ± 0%       213 ± 0%   ~     (all equal)
NewChunkMergeIterator_CreateAndIterate/chunks:_100_samples_per_chunk:_100_duplication_factor:_3_encoding:_PrometheusXorChunk-10        623 ± 0%       623 ± 0%   ~     (all equal)
NewChunkMergeIterator_CreateAndIterate/chunks:_1_samples_per_chunk:_100_duplication_factor:_1_encoding:_PrometheusXorChunk-10         15.0 ± 0%      15.0 ± 0%   ~     (all equal)
NewChunkMergeIterator_CreateAndIterate/chunks:_1_samples_per_chunk:_100_duplication_factor:_3_encoding:_PrometheusXorChunk-10         29.0 ± 0%      29.0 ± 0%   ~     (all equal)
```

```
name                                                              old time/op    new time/op    delta
NewChunkMergeIterator_Seek/scrapeInterval_30s_seekStep:_15s-10      10.1ms ± 4%    11.5ms ± 3%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_Seek/scrapeInterval_30s_seekStep:_30s-10      9.66ms ± 0%   11.09ms ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_Seek/scrapeInterval_30s_seekStep:_60s-10      9.57ms ± 0%   10.97ms ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_Seek/scrapeInterval_30s_seekStep:_300s-10     11.0ms ± 2%    12.7ms ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_Seek/scrapeInterval_30s_seekStep:_900s-10     10.2ms ± 0%    11.7ms ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_Seek/scrapeInterval_30s_seekStep:_1500s-10    8.70ms ± 3%    9.85ms ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_Seek/scrapeInterval_30s_seekStep:_3000s-10    4.13ms ± 1%    4.60ms ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_Seek/scrapeInterval_30s_seekStep:_6000s-10    1.50ms ± 1%    1.61ms ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_Seek/scrapeInterval_10s_seekStep:_5s-10       9.92ms ± 0%   11.40ms ± 1%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_Seek/scrapeInterval_10s_seekStep:_10s-10      9.68ms ± 0%   11.27ms ± 1%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_Seek/scrapeInterval_10s_seekStep:_20s-10      9.63ms ± 0%   11.11ms ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_Seek/scrapeInterval_10s_seekStep:_100s-10     10.8ms ± 0%    12.7ms ± 1%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_Seek/scrapeInterval_10s_seekStep:_300s-10     9.93ms ± 0%   11.75ms ± 1%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_Seek/scrapeInterval_10s_seekStep:_500s-10     8.47ms ± 0%    9.91ms ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_Seek/scrapeInterval_10s_seekStep:_1000s-10    4.09ms ± 0%    4.63ms ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_Seek/scrapeInterval_10s_seekStep:_2000s-10    1.49ms ± 0%    1.62ms ± 0%   ~     (p=0.100 n=3+3)

name                                                              old alloc/op   new alloc/op   delta
NewChunkMergeIterator_Seek/scrapeInterval_30s_seekStep:_15s-10       457kB ± 0%     459kB ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_Seek/scrapeInterval_30s_seekStep:_30s-10       457kB ± 0%     459kB ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_Seek/scrapeInterval_30s_seekStep:_60s-10       457kB ± 0%     459kB ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_Seek/scrapeInterval_30s_seekStep:_300s-10      457kB ± 0%     459kB ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_Seek/scrapeInterval_30s_seekStep:_900s-10      457kB ± 0%     459kB ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_Seek/scrapeInterval_30s_seekStep:_1500s-10     457kB ± 0%     459kB ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_Seek/scrapeInterval_30s_seekStep:_3000s-10     457kB ± 0%     459kB ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_Seek/scrapeInterval_30s_seekStep:_6000s-10     457kB ± 0%     459kB ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_Seek/scrapeInterval_10s_seekStep:_5s-10        457kB ± 0%     459kB ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_Seek/scrapeInterval_10s_seekStep:_10s-10       457kB ± 0%     459kB ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_Seek/scrapeInterval_10s_seekStep:_20s-10       457kB ± 0%     459kB ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_Seek/scrapeInterval_10s_seekStep:_100s-10      457kB ± 0%     459kB ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_Seek/scrapeInterval_10s_seekStep:_300s-10      457kB ± 0%     459kB ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_Seek/scrapeInterval_10s_seekStep:_500s-10      457kB ± 0%     459kB ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_Seek/scrapeInterval_10s_seekStep:_1000s-10     457kB ± 0%     459kB ± 0%   ~     (p=0.100 n=3+3)
NewChunkMergeIterator_Seek/scrapeInterval_10s_seekStep:_2000s-10     457kB ± 0%     459kB ± 0%   ~     (p=0.100 n=3+3)

name                                                              old allocs/op  new allocs/op  delta
NewChunkMergeIterator_Seek/scrapeInterval_30s_seekStep:_15s-10       6.02k ± 0%     6.02k ± 0%   ~     (all equal)
NewChunkMergeIterator_Seek/scrapeInterval_30s_seekStep:_30s-10       6.02k ± 0%     6.02k ± 0%   ~     (all equal)
NewChunkMergeIterator_Seek/scrapeInterval_30s_seekStep:_60s-10       6.02k ± 0%     6.02k ± 0%   ~     (all equal)
NewChunkMergeIterator_Seek/scrapeInterval_30s_seekStep:_300s-10      6.02k ± 0%     6.02k ± 0%   ~     (all equal)
NewChunkMergeIterator_Seek/scrapeInterval_30s_seekStep:_900s-10      6.02k ± 0%     6.02k ± 0%   ~     (all equal)
NewChunkMergeIterator_Seek/scrapeInterval_30s_seekStep:_1500s-10     6.02k ± 0%     6.02k ± 0%   ~     (all equal)
NewChunkMergeIterator_Seek/scrapeInterval_30s_seekStep:_3000s-10     6.02k ± 0%     6.02k ± 0%   ~     (all equal)
NewChunkMergeIterator_Seek/scrapeInterval_30s_seekStep:_6000s-10     6.02k ± 0%     6.02k ± 0%   ~     (all equal)
NewChunkMergeIterator_Seek/scrapeInterval_10s_seekStep:_5s-10        6.02k ± 0%     6.02k ± 0%   ~     (all equal)
NewChunkMergeIterator_Seek/scrapeInterval_10s_seekStep:_10s-10       6.02k ± 0%     6.02k ± 0%   ~     (all equal)
NewChunkMergeIterator_Seek/scrapeInterval_10s_seekStep:_20s-10       6.02k ± 0%     6.02k ± 0%   ~     (all equal)
NewChunkMergeIterator_Seek/scrapeInterval_10s_seekStep:_100s-10      6.02k ± 0%     6.02k ± 0%   ~     (all equal)
NewChunkMergeIterator_Seek/scrapeInterval_10s_seekStep:_300s-10      6.02k ± 0%     6.02k ± 0%   ~     (all equal)
NewChunkMergeIterator_Seek/scrapeInterval_10s_seekStep:_500s-10      6.02k ± 0%     6.02k ± 0%   ~     (all equal)
NewChunkMergeIterator_Seek/scrapeInterval_10s_seekStep:_1000s-10     6.02k ± 0%     6.02k ± 0%   ~     (all equal)
NewChunkMergeIterator_Seek/scrapeInterval_10s_seekStep:_2000s-10     6.02k ± 0%     6.02k ± 0%   ~     (all equal)
```